34: Create social links component to use on under construction page as well as new footer
