+++ src/components/Footer.tsx
@@ -1,16 +1,17 @@
 import { SanityDocument } from "next-sanity";
 import Link from "next/link";
-import { sanityFetch } from "../../sanity/lib/client";
+import { querySanity } from "../../sanity/lib/client";
+import SocialLinks from "./SocialLinks";
 
 const FOOTER_LINKS_QUERY = `*[_type == "link" && "footer" in category[]] | order(_createdAt asc)`;
 
 const Footer = async () => {
-	const footerLinks: SanityDocument[] = await sanityFetch<SanityDocument[]>({
-		query: FOOTER_LINKS_QUERY,
-	});
+	const footerLinks: SanityDocument[] =
+		await querySanity<SanityDocument[]>(FOOTER_LINKS_QUERY);
 	return (
 		<footer className="py-8 text-center">
-			<div className="container">
+			<div className="container flex flex-col justify-center">
+				<SocialLinks />
 				{footerLinks && (
 					<div className="hidden text-center md:flex flex-col md:flex-row justify-center gap-6 mb-4">
 						{footerLinks.map((link: any) => {
@@ -32,6 +33,7 @@ const Footer = async () => {
 					Copyright © {new Date().getFullYear()}. Henderson Force. All Rights
 					Reserved.
 				</p>
+
 				<img
 					src="/hforce_icon.svg"
 					alt="Henderson Force Icon"
